pyproject.toml
[project]
name = "online-texnika-shop"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "argon2-cffi>=25.1.0",
    "celery>=5.5.3",
    "django>=5.2.7",
    "django-ckeditor-5>=0.2.18",
    "django-jazzmin>=3.0.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "twilio>=9.8.5",
]

[tool.isort]

profile = "hug"
skip = "build,.tox,.venv"
src_paths="isort,test"